fix(adr-061): keep extraction out of frontends and policy out of model
Two Codex P1 findings on PR #994, both correct.

1. `_git_changed_paths` shells out to `git diff` to discover changed paths.
   That is a source-fact extraction routine, and `abicheck/frontends/
   AGENTS.md` states plainly that frontend code "must not contain an
   extraction routine". The function was already in `cli_scan.py` before this
   PR, but moving it into `frontends/cli/` relocated a contract violation
   into the migrated package rather than leaving it where the contract is not
   yet enforced -- which is worse than not moving it. It and its sole caller
   `_resolve_changed_seed` go back to `cli_scan.py`. Routing git discovery
   through a typed extraction entry point is the right fix and is its own
   change, not a drive-by inside a file-size slice.

2. Classifying all of `stable_abi.py` as `model` -- done so the frontend
   could reach `parse_abi3_version` -- dragged `classify`,
   `is_private_symbol` and `is_nonstable_cpython_import` into the innermost
   ring with it. Those are compatibility policy, and making them an inward
   model dependency would have weakened the boundary for every future
   consumer, not just this caller. The module genuinely holds two things, so
   only the value parser moves: `model/abi3_version.py` now owns
   `parse_abi3_version`/`format_version`/`_MAX_ABI3_MINOR` (deriving its own
   ceiling from the already-`model` `stable_abi_data`), the classification
   half stays in `stable_abi.py`, and that module's `model` classification is
   reverted. `stable_abi.py` re-exports both names, so every existing caller
   is unaffected.

`cli_scan.py` is 1772 rather than 1720 as a result -- the smaller number came
from having moved something that should not have moved. Still off the cap it
was pinned against.

Verified: `check_architecture.py` 0 findings, `ruff` and `mypy abicheck/`
clean (574 files), 1968 scan/ABI3/appcompat/reporter tests pass.
