+++ abicheck/cli_scan.py
@@ -47,6 +47,7 @@ side-effect at the bottom of :mod:`abicheck.cli` so ``@main.command`` runs.
 from __future__ import annotations
 
 import json
+import subprocess
 import sys
 import time
 from collections.abc import Callable
@@ -139,12 +140,10 @@ from .frontends.cli.scan_artifact_set import (  # noqa: F401 - re-exported under
 # still rebinds the global this module's own callers look up.
 from .frontends.cli.scan_inputs import (  # noqa: F401 - re-exported under original names
     _DURATION_UNITS,
-    _git_changed_paths,
     _normalize_depth_inputs,
     _parse_abi3_floor,
     _parse_budget,
     _resolve_auto_source_method,
-    _resolve_changed_seed,
     _scan_explicit_flags,
 )
 
@@ -622,6 +621,59 @@ def _run_artifact_set(
         click.echo(text)
     if result.exit_code != 0:
         sys.exit(result.exit_code)
+
+def _git_changed_paths(since: str, cwd: Path | None) -> list[str] | None:
+    """Paths changed vs. a git ref via ``git diff --name-only`` (no shell).
+
+    Returns the changed-path list on success (possibly **empty** for a no-op
+    diff), or ``None`` when the seed could not be produced (missing git / non-repo
+    / bad ref). The caller distinguishes the two: a successful empty diff is a
+    valid "nothing changed" seed (auto → s0), whereas ``None`` means no seed and
+    auto falls back to the mode preset (ADR-035 D7 / Codex review).
+    """
+    try:
+        proc = subprocess.run(
+            ["git", "diff", "--name-only", f"{since}...HEAD"],
+            cwd=str(cwd) if cwd is not None else None,
+            capture_output=True,
+            text=True,
+            timeout=60,
+            check=False,
+        )
+    except (OSError, subprocess.SubprocessError) as exc:
+        click.echo(f"warning: --since: could not run git diff: {exc}", err=True)
+        return None
+    if proc.returncode != 0:
+        click.echo(
+            f"warning: --since {since!r}: git diff failed "
+            f"({proc.stderr.strip() or 'non-zero exit'}); scanning broadly.",
+            err=True,
+        )
+        return None
+    return [ln.strip() for ln in proc.stdout.splitlines() if ln.strip()]
+
+
+def _resolve_changed_seed(
+    changed_paths_opt: tuple[str, ...],
+    since: str | None,
+    sources: Path | None,
+) -> tuple[list[str], str, bool]:
+    """Resolve the changed-path seed → ``(changed, changed_src, seeded)``.
+
+    ``--changed-path`` wins; else ``--since`` via git; else none. ``seeded`` tracks
+    whether a *valid* seed was produced — a successful empty diff (seeded, no
+    paths) is distinct from a missing/failed seed (not seeded): the former lets
+    auto pick s0 (no-op PR), the latter falls back to the broad mode preset
+    (ADR-035 D7 / Codex review).
+    """
+    if changed_paths_opt:
+        return list(changed_paths_opt), "--changed-path", True
+    if since:
+        git_changed = _git_changed_paths(since, sources)
+        if git_changed is None:
+            return [], f"--since {since} (seed failed; broad scope)", False
+        return git_changed, f"--since {since}", True
+    return [], "none (no diff seed; broad scope)", False
 
 _EXIT_BUDGET_OVERFLOW = 5
