+++ abicheck/frontends/cli/scan_inputs.py
@@ -20,7 +20,6 @@ exactly as before.
 
 from __future__ import annotations
 
-import subprocess
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
@@ -65,37 +64,6 @@ def _parse_budget(value: str | None) -> float | None:
     return amount * unit
 
 
-def _git_changed_paths(since: str, cwd: Path | None) -> list[str] | None:
-    """Paths changed vs. a git ref via ``git diff --name-only`` (no shell).
-
-    Returns the changed-path list on success (possibly **empty** for a no-op
-    diff), or ``None`` when the seed could not be produced (missing git / non-repo
-    / bad ref). The caller distinguishes the two: a successful empty diff is a
-    valid "nothing changed" seed (auto → s0), whereas ``None`` means no seed and
-    auto falls back to the mode preset (ADR-035 D7 / Codex review).
-    """
-    try:
-        proc = subprocess.run(
-            ["git", "diff", "--name-only", f"{since}...HEAD"],
-            cwd=str(cwd) if cwd is not None else None,
-            capture_output=True,
-            text=True,
-            timeout=60,
-            check=False,
-        )
-    except (OSError, subprocess.SubprocessError) as exc:
-        click.echo(f"warning: --since: could not run git diff: {exc}", err=True)
-        return None
-    if proc.returncode != 0:
-        click.echo(
-            f"warning: --since {since!r}: git diff failed "
-            f"({proc.stderr.strip() or 'non-zero exit'}); scanning broadly.",
-            err=True,
-        )
-        return None
-    return [ln.strip() for ln in proc.stdout.splitlines() if ln.strip()]
-
-
 def _normalize_depth_inputs(
     depth: EvidenceDepth,
     headers: tuple[Path, ...],
@@ -109,29 +77,6 @@ def _normalize_depth_inputs(
     return (), (), None, None
 
 
-def _resolve_changed_seed(
-    changed_paths_opt: tuple[str, ...],
-    since: str | None,
-    sources: Path | None,
-) -> tuple[list[str], str, bool]:
-    """Resolve the changed-path seed → ``(changed, changed_src, seeded)``.
-
-    ``--changed-path`` wins; else ``--since`` via git; else none. ``seeded`` tracks
-    whether a *valid* seed was produced — a successful empty diff (seeded, no
-    paths) is distinct from a missing/failed seed (not seeded): the former lets
-    auto pick s0 (no-op PR), the latter falls back to the broad mode preset
-    (ADR-035 D7 / Codex review).
-    """
-    if changed_paths_opt:
-        return list(changed_paths_opt), "--changed-path", True
-    if since:
-        git_changed = _git_changed_paths(since, sources)
-        if git_changed is None:
-            return [], f"--since {since} (seed failed; broad scope)", False
-        return git_changed, f"--since {since}", True
-    return [], "none (no diff seed; broad scope)", False
-
-
 def _parse_abi3_floor(abi3: str | None) -> tuple[int, int] | None:
     """Parse the --abi3 target ``Py_LIMITED_API`` floor, or ``None`` when off.
 
@@ -140,7 +85,7 @@ def _parse_abi3_floor(abi3: str | None) -> tuple[int, int] | None:
     """
     if abi3 is None:
         return None
-    from ...stable_abi import parse_abi3_version
+    from ...model.abi3_version import parse_abi3_version
 
     floor = parse_abi3_version(abi3)
     if floor is None:
